added more_vert icon for fetching recommendations and sign out and started working on front page recommendation
