+++ linkify/lib/widgets/uis/screens/artist_profile/cubit/artist_profile_state.dart
@@ -1,31 +1,31 @@
-part of 'artist_profile_cubit.dart';
+// part of 'artist_profile_cubit.dart';
 
-class ArtistProfileState {
-  final UserModel user;
-  final LoadPage status;
-  final List<SongModel> songs;
-  ArtistProfileState({
-    required this.user,
-    required this.status,
-    required this.songs,
-  });
-  factory ArtistProfileState.initial() {
-    return ArtistProfileState(
-      user: UserModel(),
-      songs: [],
-      status: LoadPage.initial,
-    );
-  }
+// class ArtistProfileState {
+//   final UserModel user;
+//   final LoadPage status;
+//   final List<SongModel> songs;
+//   ArtistProfileState({
+//     required this.user,
+//     required this.status,
+//     required this.songs,
+//   });
+//   factory ArtistProfileState.initial() {
+//     return ArtistProfileState(
+//       user: UserModel(),
+//       songs: [],
+//       status: LoadPage.initial,
+//     );
+//   }
 
-  ArtistProfileState copyWith({
-    UserModel? user,
-    LoadPage? status,
-    List<SongModel>? songs,
-  }) {
-    return ArtistProfileState(
-      user: user ?? this.user,
-      status: status ?? this.status,
-      songs: songs ?? this.songs,
-    );
-  }
-}
+//   ArtistProfileState copyWith({
+//     UserModel? user,
+//     LoadPage? status,
+//     List<SongModel>? songs,
+//   }) {
+//     return ArtistProfileState(
+//       user: user ?? this.user,
+//       status: status ?? this.status,
+//       songs: songs ?? this.songs,
+//     );
+//   }
+// }
