+++ linkify/lib/widgets/uis/screens/genre_page/cubit/genre_cubit.dart
@@ -1,32 +1,32 @@
-import 'package:bloc/bloc.dart';
-import '../../../controllers/main_controller.dart';
-import '../../../models/loading_enum.dart';
-import '../../../models/song_model.dart';
-import '../../../models/user.dart';
-import '../../../repositories/get_genre_data.dart';
+// import 'package:bloc/bloc.dart';
+// import '../../../controllers/main_controller.dart';
+// import '../../../models/loading_enum.dart';
+// import '../../../models/song_model.dart';
+// import '../../../models/user.dart';
+// import '../../../repositories/get_genre_data.dart';
 
-part 'genre_state.dart';
+// part 'genre_state.dart';
 
-class GenreCubit extends Cubit<GenreState> {
-  final repo = GenreRepository();
+// class GenreCubit extends Cubit<GenreState> {
+//   final repo = GenreRepository();
 
-  GenreCubit() : super(GenreState.initial());
-  void init(String tag) async {
-    try {
-      emit(state.copyWith(status: LoadPage.loading));
-      // var users = await repo.getUsers(tag);
-      var songs = await repo.getSongs(tag);
-      emit(state.copyWith(
-        status: LoadPage.loaded,
-        // users: users,
-        songs: songs,
-      ));
-    } catch (e) {
-      emit(state.copyWith(status: LoadPage.error));
-    }
-  }
+//   GenreCubit() : super(GenreState.initial());
+//   void init(String tag) async {
+//     try {
+//       emit(state.copyWith(status: LoadPage.loading));
+//       // var users = await repo.getUsers(tag);
+//       var songs = await repo.getSongs(tag);
+//       emit(state.copyWith(
+//         status: LoadPage.loaded,
+//         // users: users,
+//         songs: songs,
+//       ));
+//     } catch (e) {
+//       emit(state.copyWith(status: LoadPage.error));
+//     }
+//   }
 
-  void playSongs(MainController controller, int initial) {
-    controller.playSong(controller.convertToAudio(state.songs), initial);
-  }
-}
+//   void playSongs(MainController controller, int initial) {
+//     controller.playSong(controller.convertToAudio(state.songs), initial);
+//   }
+// }
