+++ linkify/lib/widgets/uis/screens/genre_page/cubit/genre_state.dart
@@ -1,27 +1,27 @@
-part of 'genre_cubit.dart';
+// part of 'genre_cubit.dart';
 
-class GenreState {
-  final LoadPage status;
-  final List<SongModel> songs;
-  // final List<User> users;
-  GenreState({
-    required this.status,
-    required this.songs,
-    // required this.users,
-  });
+// class GenreState {
+//   final LoadPage status;
+//   final List<SongModel> songs;
+//   // final List<User> users;
+//   GenreState({
+//     required this.status,
+//     required this.songs,
+//     // required this.users,
+//   });
 
-  factory GenreState.initial() =>
-      GenreState(songs: [], status: LoadPage.initial);
+//   factory GenreState.initial() =>
+//       GenreState(songs: [], status: LoadPage.initial);
 
-  GenreState copyWith({
-    LoadPage? status,
-    List<SongModel>? songs,
-    // List<User>? users,
-  }) {
-    return GenreState(
-      status: status ?? this.status,
-      songs: songs ?? this.songs,
-      // users: users ?? this.users,
-    );
-  }
-}
+//   GenreState copyWith({
+//     LoadPage? status,
+//     List<SongModel>? songs,
+//     // List<User>? users,
+//   }) {
+//     return GenreState(
+//       status: status ?? this.status,
+//       songs: songs ?? this.songs,
+//       // users: users ?? this.users,
+//     );
+//   }
+// }
